services/embedding-svc/pyproject.toml
[project]
name = "embedding-svc"
version = "1.0.0"
description = "Embedding Service — Batch-Population von pgvector-Embeddings via sentence-transformers"
requires-python = ">=3.12"

dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-health-checking>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "protobuf>=6.31",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
